feat: PetInfo 수정프로필 버전 추가, 파일 폴더 분리, import문 수정(DURI-82)
