+++ packages/ui/src/components/Quotation/index.tsx
@@ -1,4 +1,3 @@
 export * from './RequestQuotation';
 export * from './ResponseQuotation';
 export * from './Timetable';
-export * from './PetInfo';
